chatbot/chatbots/requirements.txt
pymongo==4.3.3
flask==2.2.2
ffmpeg-python==0.2.0
gtts==2.3.1
numpy==1.24.1
transformers==4.28.1
openai==0.26.5
whisper-openai==1.0.0
pydub==0.25.1
sentencepiece==0.1.98
accelerate==0.18.0
python-socketio[client]==5.9.0